Move AutoTimer to own files
